# Conflicts:
#	src/mc/common/Globals.h
#	src/mc/deps/core/file/Path.h
#	src/mc/deps/core/math/Color.h
#	src/mc/deps/core/math/Vec2.h
#	src/mc/deps/core/math/Vec3.h
#	src/mc/deps/core/string/HashedString.h
#	src/mc/entity/components/camera/aimassist/CameraAimAssistDataRegistryComponent.h
#	src/mc/nbt/CompoundTag.h
#	src/mc/network/NetworkBlockPosition.h
#	src/mc/network/packet/ModalFormRequestPacketPayload.h
#	src/mc/network/packet/ModalFormResponsePacket.h
#	src/mc/network/packet/ServerSettingsResponsePacket.h
#	src/mc/network/packet/ServerSettingsResponsePacketPayload.h
#	src/mc/network/packet/TransferPacketPayload.h
#	src/mc/platform/CallStack.h
#	src/mc/platform/threading/Mutex.h
#	src/mc/safety/RedactableString.h
#	src/mc/server/commands/CommandOutputMessage.h
#	src/mc/server/sim/MovementIntent.h
#	src/mc/world/actor/Actor.h
#	src/mc/world/item/ItemInstance.h
#	src/mc/world/item/ItemStack.h
#	src/mc/world/item/ItemStackBase.h
#	src/mc/world/item/components/LegacyOnCompleteTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnConsumeTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnDigTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnHitActorTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnHitBlockTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnHurtActorTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnUseOnTriggerItemComponent.h
#	src/mc/world/item/components/LegacyOnUseTriggerItemComponent.h
#	src/mc/world/level/BlockPos.h
#	src/mc/world/level/PlayerDimensionTransferManager.h
#	src/mc/world/level/block/BlockType.h
#	src/mc/world/level/block/actor/ChestBlockActor.h
#	src/mc/world/level/chunk/ChunkLoadActionList.h
#	src/mc/world/level/levelgen/structure/BoundingBox.h
#	src/mc/world/phys/HitResult.h

+++ src/mc/world/level/block/BurnOdds.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class BurnOdds : int {
+enum class BurnOdds : ushort {
     Instant = 100,
     Easy    = 60,
     Medium  = 20,
